+++ pkg/analyzer/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.7.0-dev
 * Updated `ConstructorElement.displayName` to either `Class` or `Class.constructor`.
 * Deprecated `InterfaceType.getSmartLeastUpperBound`, use `TypeSystem.leastUpperBound` instead.
+* Added `createMockSdk`, use it instead of `MockSdk`.
 
 ## 2.6.0
 * Deprecated `AnalysisResult.state`, check for specific valid or invalid subtypes.